do not configure Apache on Gentoo
Building snippets for make.conf here triggers cyclic dependencies when
the main portage configuration is in a different run stage.

Unfortunately it is not possible to call `package_use` multiple times
with the current version of puppet-portage, making it impossible to work
around this limitation for now (otherwise we could have USE_EXPANDed
the variables manually here and enabled them via USE flags).